requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.1.5
django-apscheduler==0.5.2
django-filter==2.4.0
django-jazzmin==2.4.7
djangorestframework==3.12.2
idna==3.2
importlib-metadata==3.4.0
Markdown==3.3.3
passlib==1.7.4
Pillow==8.1.0
pytz==2020.5
razorpay==1.2.0
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.6
zipp==3.4.0
